requirements.txt
moviepy>=1.0.3
pydub>=0.25.1
ffmpeg-python>=0.2.0
openai-whisper>=20240930
pyannote.audio>=3.1.1
torch>=2.0.0
torchaudio>=2.0.0

# LLM Post-Processing (Optional - install only what you need)
# For local processing:
# ollama

# For cloud APIs:
# openai>=1.0.0
# google-generativeai>=0.3.0
# anthropic>=0.7.0

# Search & Semantic Search (Optional - for better search capabilities)
# sentence-transformers>=2.2.0
# numpy>=1.24.0
